+++ plugin/client.go
@@ -17,3 +17,6 @@ func (p *AbyleBotterPlugin) SetAPI(api API) {
 
 // OnPost in its default implementation
 func (p *AbyleBotterPlugin) OnPost(post model.Post) {}
+
+// OnRun in its default implementation
+func (p *AbyleBotterPlugin) OnRun() {}