+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Implemented enhancements:**
 
 - Added GetVersion() to the Bot/Plugin API.
+- Added OnRun() hook from Bot to Plugin which is called when the Bot is ready to serve the plugins.
 
 **New plugins:**